requirements.txt
tortoise-orm[asyncpg]==0.24.0
aiogram==3.17.0
python-dotenv